Correct RENDER_PANELS functionality and when enabled disable HistoryPanel.
The render panels setting used to control whether the toolbar would keep
the contents of the panels in the background or render the toolbar on
every page. The history panel relies on loading panels in the background.
If panels should be rendered on the request, then the history panel is
can't/shouldn't work.